Support runtime argv in direct-native CLI programs
Lower the standard CLI args fixture through Cranelift at runtime and prove one built binary can be run with different argv values. Keep the host lowering module below the compiler source decomposition ratchet.

Refs #1477
